backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
chromadb
pydantic
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client